DiffGenerator separated from SchemaChecker
To be better testable without using container

Test stubs separated into multiple files
