+++ src/datalakebundle/table/schema/DiffGeneratorTest_stubs/df_schema.py
@@ -1,0 +1,24 @@
+import pyspark.sql.types as t
+
+df_schema = t.StructType(
+    [
+        t.StructField("FIELD1", t.IntegerType()),
+        t.StructField("FIELD2", t.StringType()),
+        t.StructField(
+            "STRUCT1",
+            t.StructType(
+                [
+                    t.StructField("NESTED_FIELD1", t.StringType()),
+                    t.StructField(
+                        "STRUCT2",
+                        t.StructType(
+                            [
+                                t.StructField("NESTED_FIELD2", t.StringType()),
+                            ],
+                        ),
+                    ),
+                ],
+            ),
+        ),
+    ],
+)
