+++ src/datalakebundle/table/schema/DiffGeneratorTest_stubs/df_schema3.py
@@ -1,0 +1,34 @@
+import pyspark.sql.types as t
+
+df_schema3 = t.StructType(
+    [
+        t.StructField("FIELD1", t.IntegerType()),
+        t.StructField("FIELD2", t.StringType()),
+        t.StructField("FIELD3", t.StringType()),
+        t.StructField(
+            "STRUCT1",
+            t.StructType(
+                [
+                    t.StructField("NESTED_FIELD1", t.StringType()),
+                    t.StructField(
+                        "STRUCT2",
+                        t.StructType(
+                            [
+                                t.StructField("NESTED_FIELD2", t.StringType()),
+                            ],
+                        ),
+                    ),
+                    t.StructField(
+                        "STRUCT3",
+                        t.StructType(
+                            [
+                                t.StructField("NESTED_FIELD3", t.StringType()),
+                                t.StructField("NESTED_FIELD4", t.StringType()),
+                            ],
+                        ),
+                    ),
+                ],
+            ),
+        ),
+    ],
+)
