+++ src/datalakebundle/table/schema/DiffGeneratorTest_stubs/df_schema4.py
@@ -1,0 +1,7 @@
+import pyspark.sql.types as t
+
+df_schema4 = t.StructType(
+    [
+        t.StructField("FIELD1", t.IntegerType()),
+    ]
+)
