+++ src/datalakebundle/table/schema/DiffGeneratorTest_stubs/expected_schema2.py
@@ -1,0 +1,24 @@
+import pyspark.sql.types as t
+
+expected_schema2 = t.StructType(
+    [
+        t.StructField("FIELD1", t.IntegerType()),
+        t.StructField("FIELD2", t.StringType()),
+        t.StructField(
+            "STRUCT1",
+            t.StructType(
+                [
+                    t.StructField("NESTED_FIELD1", t.DateType()),
+                    t.StructField(
+                        "STRUCT2",
+                        t.StructType(
+                            [
+                                t.StructField("NESTED_FIELD2", t.StringType()),
+                            ],
+                        ),
+                    ),
+                ],
+            ),
+        ),
+    ],
+)
