+++ src/datalakebundle/table/schema/DiffGeneratorTest_stubs/expected_schema5.py
@@ -1,0 +1,27 @@
+import pyspark.sql.types as t
+
+expected_schema5 = t.StructType(
+    [
+        t.StructField("FIELD1", t.IntegerType()),
+        t.StructField("FIELD2", t.StringType()),
+        t.StructField("FIELD3", t.StringType()),
+        t.StructField(
+            "STRUCT1",
+            t.StructType(
+                [
+                    t.StructField("NESTED_FIELD1", t.StringType()),
+                    t.StructField(
+                        "STRUCT2",
+                        t.StructType(
+                            [
+                                t.StructField("NESTED_FIELD0", t.StringType()),
+                                t.StructField("NESTED_FIELD00", t.StringType()),
+                                t.StructField("NESTED_FIELD2", t.StringType()),
+                            ],
+                        ),
+                    ),
+                ],
+            ),
+        ),
+    ],
+)
