+++ src/datalakebundle/table/schema/DiffGeneratorTest_stubs/expected_schema_array.py
@@ -1,0 +1,12 @@
+import pyspark.sql.types as t
+
+expected_schema_array = t.StructType(
+    [
+        t.StructField("FIELD1", t.IntegerType()),
+        t.StructField("FIELD2", t.DoubleType()),
+        t.StructField(
+            "ARRAY1",
+            t.ArrayType(t.ArrayType(t.StringType())),
+        ),
+    ],
+)
